[brain tests] Burst multiprocessing from the main file
